pyproject.toml
[tool.poetry]
name = "mapillary_proxy"
version = "0.1.0"
description = ""
authors = ["Oleg Lavrovsky <oleg@datalets.ch>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
falcon-caching = "^1.1.0"
falcon = "^3.1.0"
requests = "^2.28.0"
python-dotenv = "^0.20.0"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
